Prepare file check dependencies for several BPF targets

+++ qemu-csd/bpf_helpers/include/bpf_helpers_prog.h
@@ -7,12 +7,12 @@
  */
 
 // Called by BPF to allow copying return data out of running BPF program.
-static void (*bpf_return_data)(void *data, size_t size) = (void *) 1;
+static void (*bpf_return_data)(void *data, uint64_t size) = (void *) 1;
 
 // Called by BPF to perform an 'on device' read at a specific LBA.
 static void (*bpf_read)(uint64_t lba, void *data) = (void *) 2;
 
 // Called by BPF to get the size in bytes per LBA.
-static size_t (*bpf_get_lba_siza)(void) = (void *) 3;
+static uint64_t (*bpf_get_lba_siza)(void) = (void *) 3;
 
 #endif //QEMU_CSD_BPF_HELPERS_PROG_H
